Extract private channel indexing helper
